requirements.txt
starlette==0.12.9
uvicorn==0.10.9
pydantic==1.2
fastapi==0.45.0
python-dotenv==0.10.3
gunicorn